+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add show and hide Minimize label (#144)
 - Added maxVisibleValues option in minimize view (#147)
 - Add plugin e2e tests and remove cypress (#149)
+- Add Row count in Table view (#148)
 
 ## 2.5.0 (2024-04-01)