+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 - Update selecting and deselecting multi-levels (#139)
 - Update Auto/Static Minimize label width (#144)
 - Add show and hide Minimize label (#144)
-- Added maxVisibleValues option in minimize view (#147)
+- Added maximum visible values option in minimize view (#147)
 - Add plugin e2e tests and remove cypress (#149)
 - Add Row count in Table view (#148)
+- Update AutoScroll Behavior (#152)
 
 ## 2.5.0 (2024-04-01)